+++ ExilenceNextApp/src/store/settingStore.ts
@@ -7,7 +7,7 @@ import { RootStore } from './rootStore';
 
 export class SettingStore {
   @persist @observable lowConfidencePricing: boolean = false;
-  @persist @observable autoSnapshotting: boolean = false;
+  @persist @observable autoSnapshotting: boolean = true;
   @persist @observable priceTreshold: number = 0;
   @persist @observable autoSnapshotInterval: number = 60 * 2 * 1000; // default to 2 minutes
   @persist